+++ src/components/index.ts
@@ -24,16 +24,8 @@ export { Form } from './Form';
 export { FormItem } from './FormItem';
 export { GroupJoinForm } from './GroupJoinForm';
 export { Header } from './Header';
-export { HotKeyRecorder, useHotKey } from './HotKey';
+export { useHotKey } from './HotKey';
 export { ImageOCRProgress } from './ImageOCRProgress';
-export { ImageSelect } from './ImageSelect';
-export { ImageSourceViewer } from './ImageSourceViewer';
-export { ImageSourceViewerGod } from './ImageSourceViewerGod';
-export { ImageSourceViewerModeControl } from './ImageSourceViewerModeControl';
-export { ImageSourceViewerProofreader } from './ImageSourceViewerProofreader';
-export { ImageSourceViewerTranslator } from './ImageSourceViewerTranslator';
-export { ImageTranslatorSettingHotKey } from './ImageTranslatorSettingHotKey';
-export { ImageTranslatorSettingMouse } from './ImageTranslatorSettingMouse';
 export { ImageViewer } from './project-file/ImageViewer';
 export { InvitationList } from './InvitationList';
 export { InviteUser } from './InviteUser';
@@ -69,7 +61,6 @@ export { TeamSearchList } from './TeamSearchList';
 export { TeamSettingBase } from './TeamSettingBase';
 export { Tooltip } from './Tooltip';
 export { TranslationProgress } from './TranslationProgress';
-export { TranslationSaveFailed } from './TranslationSaveFailed';
 export { TranslationUser } from './TranslationUser';
 export { TypeRadioGroup } from './TypeRadioGroup';
 export { UserEmailEditForm } from './UserEmailEditForm';
