+++ src/components/project-file/ImageSourceViewerGod/Translation.tsx
@@ -6,19 +6,19 @@ import { lighten } from 'polished';
 import React, { useEffect, useState } from 'react';
 import { useIntl } from 'react-intl';
 import { useDispatch, useSelector } from 'react-redux';
-import { DebounceStatus, Icon, Tooltip, TranslationUser } from '..';
-import { PROJECT_PERMISSION } from '../../constants';
-import { FC, Source } from '../../interfaces';
-import { Translation as ITranslation } from '../../interfaces/translation';
-import { AppState } from '../../store';
+import { DebounceStatus, Icon, Tooltip, TranslationUser } from '@/components';
+import { PROJECT_PERMISSION } from '@/constants';
+import { FC, Source } from '@/interfaces';
+import { Translation as ITranslation } from '@/interfaces/translation';
+import { AppState } from '@/store';
 import {
   editMyTranslationSaga,
   editProofreadSaga,
   selectTranslationSaga,
-} from '../../store/source/slice';
-import style from '../../style';
-import { clearClickEffect, clickEffect } from '../../utils/style';
-import { can } from '../../utils/user';
+} from '@/store/source/slice';
+import style from '@/style';
+import { clearClickEffect, clickEffect } from '@/utils/style';
+import { can } from '@/utils/user';
 
 export interface OnTextAreaChange {
   (event: React.ChangeEvent<HTMLTextAreaElement>): void;
