+++ src/components/project-file/ImageTranslatorSettingMouse.tsx
@@ -1,7 +1,7 @@
 import { css } from '@emotion/core';
 import React from 'react';
 import { useIntl } from 'react-intl';
-import { FC } from '../interfaces';
+import { FC } from '@/interfaces';
 import classNames from 'classnames';
 
 /** 模板的属性接口 */
